Add files via upload
CARD is the 'Aadhar card' that is being scanned into the OCR code. 
The card is converted to black and white, then denoising is performed followed by image sharpening.
Thus, we obtain the correct AAdhar card number as the output.

Doubt: How to extract only numeric data from the Aadhar card?